--- chain/src/main/java/io/nuls/chain/model/Seed.java
@@ -1,27 +1,0 @@
-package io.nuls.chain.model;
-
-/**
- * @author tangyi
- * @date 2018/11/7
- * @description
- */
-public class Seed {
-    private String ip;
-    private int port;
-
-    public String getIp() {
-        return ip;
-    }
-
-    public void setIp(String ip) {
-        this.ip = ip;
-    }
-
-    public int getPort() {
-        return port;
-    }
-
-    public void setPort(int port) {
-        this.port = port;
-    }
-}
